Turn off DNS creation
Running into problems where the zone that it attempts to create already
exists, even if it doesn't exist at the beginning of terraform apply.
This seems to indicate a double creation somewhere in our templates.

We don't use these DNS entries currently, so not creating them should be
fine for now.

[#178007284] Fix AWS terraform in Ops Manager CI

Signed-off-by: Brian Upton <bupton@vmware.com>